Move friction models into separate file; Make a model chooser function (for now)
